+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v.0.3.1
 
 * **[Breaking]** Update `round()` for `Vec` instances because julia 1.11 overload resolution made it impossible :(
+* Add `box_indices(array, I=Int)` to get a `Box` covering an array.
 
 # v0.3.0